+++ iroko/persons/fixtures.py
@@ -20,6 +20,14 @@ def _is_cuban_affiliation(affiliation: str):
             return True
     return False
 
+def _is_university_affiliation(affiliation: str):
+    fix_words = ['universidad', 'university']
+    af = normalize('NFC', affiliation.lower())
+    for word in fix_words:
+        if word in af:
+            return True
+    return False
+
 
 def _creator_is_cuban(creator):
     if 'affiliations' in creator:
@@ -49,11 +57,30 @@ def get_cuban_authors_from_record(rec: IrokoRecord):
 def get_all_cubans_authors_from_records():
     search = IrokoRecordSearch()
     cubans = dict()
+    universities = dict()
     for hit in search.scan():
         record = IrokoRecord.get_record_by_pid_value(hit.id)
         authors = get_cuban_authors_from_record(record)
         for aut in authors:
             if 'name' in aut and aut['name'] not in cubans:
                 cubans[aut['name']] = aut
-    return cubans
+                for aff in aut['affiliations']:
+                    if _is_university_affiliation(aff):
+                        universities[aut['name']] = aut
+    return cubans, universities
 
+def _tmp_func():
+    search = IrokoRecordSearch()
+    last:str = '2022-12-31'
+    universities = dict()
+    for hit in search.scan():
+        record = IrokoRecord.get_record_by_pid_value(hit.id)
+        cur = record['publication_date']
+        if last > cur:
+            last = cur
+            print('---------------------')
+            print('---------------------')
+            print(last)
+            print(record)
+            print('---------------------')
+            print('---------------------')
